refactor(Wings): prepare module for easy mode expansion
- Lines mode WIP
