+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/ModuleManager.kt
@@ -230,7 +230,7 @@ import net.ccbluex.liquidbounce.features.module.modules.render.ModuleTNTTimer
 import net.ccbluex.liquidbounce.features.module.modules.render.ModuleTracers
 import net.ccbluex.liquidbounce.features.module.modules.render.ModuleTrueSight
 import net.ccbluex.liquidbounce.features.module.modules.render.ModuleVoidESP
-import net.ccbluex.liquidbounce.features.module.modules.render.ModuleWings
+import net.ccbluex.liquidbounce.features.module.modules.render.wings.ModuleWings
 import net.ccbluex.liquidbounce.features.module.modules.render.ModuleXRay
 import net.ccbluex.liquidbounce.features.module.modules.render.ModuleZoom
 import net.ccbluex.liquidbounce.features.module.modules.render.cameraclip.ModuleCameraClip
